Ensure script runs from application directory
Previously, executing script caused pygds to throw an error related
to locating the relevant DLLs. Executing a script as admin changes
the directory from which the PowerShell script runs, so it was not
able to locat the g.tec DLLs.

Location of PSS was modified to run from application directory,
where the DLLs are located.